pc_std: Enable OTA update generation
() Correctly set BOARD_BOOTIMAGE_PARTITION_SIZE to 10M

() Add recovery.fstab for PC targets. Due to the way the build
system works, it expects the file to be in the product directory;
use symlinks to accomplish this.

() Use alternate SYSLINUX menu (which isn't hidden) and Diskinstaller
configuration if Fastboot is staged

() Add bootloader and Droidboot image (if applicable) to the
INSTALLED_RADIOIMAGE_TARGET, which in fact has nothing to do with
radios; it is a means to place extra files in an OTA target-files-package

() Implement Releasetools extensions to update the bootloader and
Droidboot images. This currently has a problem with incremental updates
(bad SHA1 match after applying a patch) which will be addressed in
a subsequent patch

() Add 'otapackage' to $(TARGET_PRODUCT) Make targets

Change-Id: I9edb0f73b75ebcd657dc4fd8aad33175930a3207
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>