turn on bugmailer
Bugmailer is a service which is invoked by a magic key. When initiated,
it collects dumpstate output and captures a screenshot, and then invokes
GMail to mail the bug report to the user's GMail inbox, where the report
can be inspected or forwarded as deemed necessary.

On Samsung tablet, the magic keys sequence is Vol+ and Vol-
On Lenovo, F1 and F12

Change-Id: I53e03170aeead2c175122283adeceaaa80c8350a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>