create libminimal_recovery_ui; use it on Samsung
The default recovery UI assumes the presence of KEY_ENTER
and KEY_HOME, neither are the case on devices like Samsung.

Create a minimal recovery UI plug-in which only assumes
the volume and power buttons. The UI is shown/hidden by
pressed Vol+ and Vol- simultaneously, and a menu item
is chosen with the power button.

Set up the Samsung recovery console to use this plug-in.
We don't really need it for Lenovo.

Change-Id: I6bd10839482c4582dc850e09d15838bff5ff1872
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>