audio_hal.c: Fix resource leak in intel_hda_open()
An allocated struct is not freed before returning with
with an error. This adds a free() operation.

Also added the call to _EXIT() before this free() so
that we know when it leaves the function.

Note that we should not free the struct if the function
succeeds. The struct is needed by audio flinger for
proper operation.

Change-Id: I21ba488f04281a01611c5cfb073476a917c8c214
Original-Change-Id: Ifeb4286fd849b8e866c174ac98030313d5398c60
Signed-off-by: Daniel Leung <daniel.leung@intel.com>