Change name of wifi driver module for 3.4 kernel
The name of the module "iwlagn" for intel AGN wifi changes
to "iwlwifi" between the 3.0 kernel and 3.4 kernel.  This
patch makes the corresponding change in the target-specific
BoardConfig files.

The selection of the module is conditional based on the
following rules:

 () If building a kernel directly, examine the iwlwifi
    module's Kconfig file to determine its module name.

 () If using a prebuilt kernel, examine its modules file for
    the name of the module.

The patch also fixes the name of the driver-arguments build
variable WIFI_DRIVER_MODULE_ARG that was wrong in (only) the
lenovox220t BoardConfig file.

Issue: AIA-439
Change-Id: I3db5e4195eebefb270aac8fa743b7325cadeecb6
Depends-Change-Id: Ic5ed929779b3450d573f288d8c189925424a7ecf
Signed-off-by: Paul Drews <paul.drews@intel.com>