Restrict wifi driver-name search for more-exact match
The search string for examining a Kconfig file for the
iwlwifi driver configuration was not sufficiently
restricted.  It allowed symbols with other characters
following the "config IWLWIFI" such as "config
IWLWIFI_DEBUG" to be erroneously recognized as configuring
the iwlwifi driver.  A consequence was that the build for
lenovox220t with 3.0 kernel erroneously concluded that the
driver should be named "iwlwifi".

This patch restricts the regular expression so that the
config symbol must be a word-end or line-end.

Change-Id: I0f3916079195278d4232155f42b5bd51dab0c67d
Signed-off-by: Paul Drews <paul.drews@intel.com>