Enable battery polling by default in overlay for Lenovo X220t
The option to enable battery polling is stored in the main system
settings database. This causes the default value for battery
polling to be set to enable when this database is created the
first time, for Lenovo X220t. For other devices, this can be
enabled in the option under Settings/Battery.

Issue: AIA-432
Change-Id: Ibc0b69b51b8677372bb3fd14d9749fcb1a54de65
Depends-Change-Id: Idf44606ba3a566341bf5952b21edce3b4101f457
Signed-off-by: Daniel Leung <daniel.leung@intel.com>