MUST_REVERT:  Workaround excessive tx retransmit with 11n
The IWLWIFI driver has been repeatedly reported to either
fail or work very slowly when 11n support is enabled (as is
default).  In a moderately noisy office environment,
i88880211 debug features were used to track transmit-retry
statistics.  It was found that these exceeded tx packets by
a factor of 10X, i.e., approximately 10 retries were
required to transmit each packet.  Wifi was unusable under
these circumstances, with the symptom that it was usually
impossible to resolve a host-name and get a connection to a
web site (by name) in the browser.  For various reasons, it
was sometimes possible to maintain a connection or resolve
subsequent names after first having contacted a web site by
IP address.

Other users on the web reported improvements from disabling
11n support in the driver.  This patch uses one of the
supported IWLWIFI module parameter settings to disable 11n.
Observed results show that this reduces the transmit-retry
rate to approximately the transmit packet rate.  On several
different systems the retry-to-tx ratio was about 1.2X to
0.8X.  WiFi becomes usable under these circumstances.  Host
names can be readily resolved and connections can be readily
established and maintained.

Note that this is still a workaround, and ongoing work to
find and fix the root cause is still being tracked in
AIA-443.

For: AIA-443
Change-Id: I1115e2556a50d15678ce3c50945d49ff5c1ccedc
Signed-off-by: Paul Drews <paul.drews@intel.com>