MUST_REVERT:  Workaround rf-reset from PLCP health checks
The IWLWIFI driver recently added a health-check for errors
detected in Physical Layer Convergence Protocol (PLCP)
headers.  When the error-rate per normalized 100ms interval
exceeds a threshold, this is considered a health-check error
and the radio is restarted.

In a moderately crowded office environment, the observed
distribution of these errors over a 1-hour test run is such
that the mean time between failures (exceeding the
health-check error threshold) at different threshold levels
was as follows:

  Errors per 100ms      Mean Time Between Failures
    50                    1.9 sec
    100                   10.3 sec
    200                   387.7 sec

At the defined threshold of 50 errors per 100ms, the radio
reset itself continuously, making wifi unusable.  Moving
this threshold to the highest predefine rate of 200 errors
per 100ms has been found work.  However, this would only be
a workaround rather than addressing the basic problem, and
there was reluctance to contaminate the kernel code with the
workaround.

Accordingly, this patch uses a module-parameter of the
iwlwifi driver to disable the plcp health-check completely
as a workaround for the problem.  This is part of several
workaround patches to make the wifi usable on the devices
while the root problem can be worked on in parallel.

The (somewhat mis-titled) AIA-443 bug report tracks the
overall progress against this problem.

For: AIA-443
Change-Id: Ieced15e80b5072b1e341eb898cb6459cc1dd7267
Signed-off-by: Paul Drews <paul.drews@intel.com>