Edit the bootanimation.zip for JB
There is a new mechanism for handling bootanimation exit in JB and
an unintended consequence of allowing the bootanimation to exit
gracefully now is that animations that define pauses between parts
might have ridiculously long waits until exit.

Our animation defined a 10 second pause between infinite cycles.
Because of what I suspect is a bug in the bootanimation code, this
was getting multiplied by the duration of the last frame draw
(which can be between a few ms to hundreds of ms depending on CPU
load/timing).  This is why we were seeing a very long pause until
the animation exited gracefully. To correct this issue, the
bootanimation was modified to remove the pause and the infinite
looping behaviour.

The second line of desc.txt (in bootanimation.zip) changed from:
p 0 10000 part0
to
p 1 0 part0

The first p now signifies that the animation does not need to play to
completion ('c' would signify that it must).  The second value is number
of iterations for that animation part (0 being infinite loop), the third
value is the pause between animation parts (since it now only plays one
time it should be 0), the last value is the path to the animation folder.

Change-Id: Iafac69179bbb6fface2b58833fc19aebe484fce5