MUST_REVERT Restore WIFI_DRIVER_MODULE_PATH
Restores the old WIFI_DRIVER_MODULE_PATH contract between
the BoardConfig and libhardware_legacy.  This is to get wifi
up and running quickly.  In the longer run we need to
reinstate the module-loading technology we had shortly
before JB.

The change leaves the file cross-compatible between the "new
way" libhardware_legacy and "old way" libhardware_legacy to
allow the libhardware_legacy change to be reinstated easily
without cross-project interdependencies.

Change-Id: Ie834bd47b9fe3d7d474dc09cf286f56818381530
Signed-off-by: Paul Drews <paul.drews@intel.com>