Update wpa_supplicant service invocation parameters
The default parameters were insufficient for the
wpa_supplicant to know what interface to use and where to
communicate with the driver regarding the control interface.
This change makes these things explicit in the
wpa_supplicant service invocation so wifi can work.

In addition, the management of the wpa_supplicant.conf file
entry for "ctrl_interface" is limited to a hard-coded value
in wpa_supplicant, so this hard-coded value is used.

Change-Id: I56dac708cfc414dc26a2c2e3f7116357de446910
Signed-off-by: Paul Drews <paul.drews@intel.com>