Copy all idc files for a target
An idc files is tied to an input device connected to the board. Once
system can create input devices based on the detection in kernel, we
can just collect all supported idc files in a single one build, and
framework will load the correct idc file for a probed input device
at run time.

For: AXIA-689
Change-Id: I7fc7a83d2c8cbfc636b444b399622d0a3c782d3b
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>