Use Dalvik profile with bigger heap size
The default Dalvik heap size for tablet devices is 256MB.
This change is to inherit the Dalvik profile for high
resolution devices, which have heap size of 384MB.

This also helps performance for apps with large memory
requirements. The heap growth size is larger so there
are fewer malloc events.

Change-Id: Ie36b04f38b4c9e8c1b6e9fcc61d9ebbd1ac4b0be
Signed-off-by: Daniel Leung <daniel.leung@intel.com>