symlink fstab.$(TARGET_PRODUCT) to fstab.common
Some parts of the build insist that the fstab file be named
fstab.$(TARGET_PRODUCT). In order to avoid having to migrate the pc_std
init.rc directives that use it to the product init.rcs, we symlink it
instead.

If a board defines BOARD_HAS_FSTAB, this link is not created.

Change-Id: I03e2d92399f357b15c1e60b1d81b2c3e6c468bfb
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>