Make a safer OTA update for bootloader
OTA update used to include the bootloader partition and dd-ing onto
the device. If power fails during update, this will cause the device
to be unable to boot up again.

The safer way is to copy the necessary files one by one onto boot
partition with different names. After the file is copied, and it will
be renamed to take over the old one. This way, there will still be
a working copy.

Issue: AXIA-404
Change-Id: Idb41f5b6b8cf67343d6d8eb3be1c8da84f61ab79
Depends-Change-Id: I86f4e40cd3d40f22f5557c2b762ee5491f4c9d2f
Signed-off-by: Daniel Leung <daniel.leung@intel.com>