Revert "Resize /data instead of /system"
This reverts commit 2e5ac2d75034d99352dc8309fd05289c99a95fbb.

() System needs to be resized to fill the partition in case of an
OTA update which has more data than the currents system image
() Data should NOT use userdata.img and should not be reaized.
Instead when the filesystem is created using mkfs, the proper
footer size should be accounted for.