bigcore/libsensors: check for presence of IIO sensors
bigcore libsensors BoardConfig was assuming IIO sensors are present on
the device. Some devices (e.g. samsung xe700t) do not have them and so
they should not be reported to Android's sensorservice.

For now, this is just a very simplistic yes/no test to support both
the xe700t as well as the new Win8 samsung device that actually has the
IIO sensors.

Change-Id: Ib920315aa02442f9f814d78ee1a496237b74e537
For: AXIA-837
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>