init scripts: start eth0 setup when eth0 is plugged in
This allows adb via USB ethernet connectivity even when the USB
ethernet adapter is plugged in late after boot (or removed and
later reinserted).

This patch also removes the sleep from the eth0 bring-up script since
now the script is called after eth0 is detected and thus the interface
is always available at the time the script is run.

It also removes the net.eth0.startonboot property since we don't need
it anymore.

Depends-Change-Id: I8f032ef983aeb36dd013fafc4a5df5238decbb17
Change-Id: I0ca96024a5877897f9340dc3accff7cfb88932e4
Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>