libsensors: save sensor presence from first boot
Because of delayed module loading and USB enumeration time, IIO sensors
sysfs nodes may not yet be created by the time libsensors is asked to
initialize by the frameworks. This patch will wait up to 10 sec for the
sysfs nodes to appear before reporting "no sensors" to the framework on
the first boot of the device and set a system property based on whether
or not the iio sensors were found.

On subsequent boots, it will only consult this system property and thus
no delay in system services startup.

Change-Id: Ibb7b6d69ce1c1dcadd0ddaa361a4297489a044c7
Signed-off-by: Matt Gumbel <matthew.k.gumbel@linux.intel.com>