New module name for Intel Wifi driver
Someone decides it is good to change the name of wifi module for
every major kernel release. So, again, the userspace needs to be
updated to load the new driver. To make matter worse, they
decided to split the module in two, where iwlwifi retains some
core functionality and iwldvm is the main module to load for
interface to show up. Therefore it is no longer possible to
send parameters directly to iwlwifi, as the wifi HAL does not
load iwlwifi directly.

Change-Id: I871847c225376a0baf8d2437ea2520bc9d12dbb7
Signed-off-by: Daniel Leung <daniel.leung@intel.com>