init.bigcore.rc: Set perms on BCB block device node
Now that the Bootloader Control Block is updated in userspace
in android_reboot(), we need to allow processes in the 'system'
group to be able to write to it. This wasn't previously necessary
because the logic for this used to be in the kernel, done in the
context of a reboot handler.

Fixes a problem where the Settings app couldn't initiate a
factory data reset because it wasn't able to update the BCB.

Issue: AXIA-1011
Change-Id: I10803eb406f55e8636881b5be915ad3f50215f30
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>