Prevent ACPI Video from restoring LCD brightness at resume
ACPI Video restores its own LCD brightness independent of what the DRM
driver (e.g. i915 driver) is doing. For the build targets using i915
driver, the light HAL is controlling the backlight in i915 but not in
the ACPI Video driver. When the device resumes, the ACPI Video restores
its own brightness value, while i915 has brightness at 0 (which is set
by frameworks). This causes weird user experience as they can see the
text console flashing.

This change tells the ACPI Video module not to touch the brightness
of LCD backlight, for targets not using the VESA framebuffer.

Change-Id: I8764afb2226b8fcd5c4c2e7bf46efde5bd1e31eb
Signed-off-by: Daniel Leung <daniel.leung@intel.com>