Camera: Enable camera app on ivb
Camera icon will disappear in app home screen by Google's design
when device reports to upper layer that it has no back camera.
This change is effective since JB release.

But because people still want it on IA devices which has only
front camera and we don't think it is justified to hack Google's
app to achieve it, we have to do this cheat by reporting front
camera as 'back camera'.

For: AXIA-1017
Change-Id: I2467edeeff0b8734ee0f78634b89c3b33f25c422
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>