Bootloader image has been replaced with dbupdate.bin
Following the latest fastboot update, the bootloader file is no longer
needed and the flashing script needs the new dbupdate.bin file. This
patch modifies the droidboot target to include this file into the
droidboot zip archive, and fixes the current breakage in the smoke
tests system, which makes all flashing code to fail.

Signed-off-by: Alin Dobre <alin.m.dobre@linux.intel.com>