sync_img: Update to new way of referencing block devices
The old nodes aren't correct in the new partitioning scheme.
This fixes it permanently.

Bootloader partition is now customized per-installation. It can't
be flashed to the device any more like this. If someone really wants
to update the bootloader they need to use Fastboot.

Removed syslinux-image from dbimages deps since nothing should be
using it.

Change-Id: I52e7f2ca5fb537c3623f03838f75a555623ba252
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>