[PATCH] audio: fix crash when starting capture
From 602db275614c9ea17f4165f2814f301fcf0f85d0 Mon Sep 17 00:00:00 2001
From: Eric Laurent <elaurent@google.com>
Date: Wed, 29 Aug 2012 18:51:31 -0700

When starting capture, we check if a playback stream is active with an
incompatible sampling rate and if yes force this stream in standby.
The output stream lock is acquired while requesting standby. The problem
is that this stream is referenced by adev->active_out which is cleared by
do_output_standby() and a crash occurs when releasing the lock.

The same problem is present when starting a playback stream and a capture
stream is active.

For: AXIA-1171

Signed-off-by: Charles Johnson <charles.f.johnson@intel.com>

Change-Id: I72ce058308371199655e3b2e1cfc0529328555ec