bigcore/wifi: Use rfkill instead of module un/loading
Conformant mainline drivers support rfkill as the proper means to
disable a wifi device.  Use that as the implementation of
wifi_un/load_driver(), and suddenly we don't need any of the module
loading logic and can simply rely on the modalias handling in ueventd.
And the code is more robust too, as a new (nl80211-compatible) wifi
device can be configured in the kernel and used automatically without
updating the HAL.

Issue: AXIA-1727
Change-Id: I90f225657d3ce29ea175783426aceed00373e25a
Signed-off-by: Andy Ross <andy.ross@windriver.com>