common: revert to AOSP's perf
external/perf/ depends on Bison 2.5. However, the bison that comes
with Android is 2.3. In external/perf/Android.mk, bison is called
using a relative path and thus will pick up the 2.3 version from out/
if it has been built, otherwise it picks up the version from the host
Linux distro. This means that perf sometimes builds and sometimes fails
to build.

For now, we downgrade to use perf from AOSP which builds fine using
Bison 2.3.

Change-Id: I111229c786bbc0de527f05ae507bb53376d89c6e
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>