bigcore: Give more time for iio device detection
Currently android framework checks if /sys/bus/iio/devices node
is created and sets sensor property based on it. The check is
performed in a for loop of ~1 sec.
Looks like for i2c based sensors it is taking more time for the
node to be created and hence setting of property fails.

Increase the loop time to ~5 sec max which gives enough time for
the node to be detected.

Issue: AXIA-2039
Change-Id: I1b33eae948482e9e6835918d61357cd6ba3d7b2b
Signed-off-by: Huzefa Kankroliwala <huzefa.nomanx.kankroliwala@intel.com>