flash-pc: don't flash bootloader with -a
Droidboot is intentionally omitted from -a because usually the user
does not want to risk "bricking" their device by breaking the flashing
mechanism. On those grounds, we would also not want to update the
bootloader by default.

This is particularly an issue for the auto-test slaves where we cannot
risk a bad droidboot or bad bootloader getting flashed and failing every
test job thereafter.

Change-Id: If3389a752c6734ca02db040db3424045dc184a7d
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>