acer-w700: swap camera order
The default media_profiles.xml (and common convention, on at least the
Galaxy Nexus) has camera 0 as the "front" camera and 1 as "back".  On
the Acer specifically, this happens to match the hardware and works
around the CTS test failure in the linked bug, and also the GMS Camera
app's inability to record in 1080p.  A real fix will require
scalability work, see bug.

Change-Id: Iaa56363f55bcc31001e7d2a995ac6cb601f112e3
For: AXIA-2223
Signed-off-by: Andy Ross <andy.ross@windriver.com>