Audio: Enable Intel HD-Audio on HSW-ULT
Modify the Audio HAL so that it probes to determine which
Intel HD-Audio controller to use for system audio.  The is required
since on HSW-ULT systems, there may be a second Intel HD-Audio controller
in the north complex which gets enumerated first.  This code probes for
the audio controller in the PCH.

Issue: AXIA-1990

Change-Id: I1bd3b466da99f5370b867b0fc58ea95abf0a647e
Signed-off-by: Charles Johnson <charles.f.johnson@intel.com>