support: regen shim test key/cert with longer subject line
The length of the binary cert in DER format depends on the length
of the subject line. This provides a challenge for us when we
replace the binary blob inside shim for production release
(for use with build/tools/releasetools/sign_target_files_apks).

This regenerates the shim key with the maximum allowable subject
line with 64 characters (as allowed by OpenSSL).

For: AXIA-2387
Change-Id: I55472ec4321ace0b6899af4e07c2ca1747e2d545
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Reviewed-on: https://otc-android.intel.com/gerrit/19032
Reviewed-by: KeithX W Thomas <keithx.w.thomas@intel.com>
Reviewed-by: Matthew K Gumbel <matthew.k.gumbel@intel.com>
Reviewed-by: Andrew P Boie <andrew.p.boie@intel.com>
Tested-by: Andrew P Boie <andrew.p.boie@intel.com>