Sign kernel modules with DB key pair in BIOS
The intention is that kernel will import DB certificates from BIOS
to verify modules at run time in the future.

However, because the current kbuild also embeds the same pub key
into kernel at build time, the actual effect so far is same as
before, except signing with a different key. For the same reason,
this change doesn't depend on the implementation of importing
BIOS keys in kernel at this time.

For: AXIA-2485

Change-Id: I9f87d72ba5e4649d5b7a5940440a0e5e8e20efb2
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Reviewed-on: https://otc-android.intel.com/gerrit/19040
Reviewed-by: Matthew K Gumbel <matthew.k.gumbel@intel.com>
Tested-by: Matthew K Gumbel <matthew.k.gumbel@intel.com>