releasetools: implement scalable capsule updates
() Update the capsule pushing code to the current version of the
kernel driver interface

() Fixup use of ErrorAbort; if used multiple times, only the last
invocation is propagated to the user. Recovery logs all stdout,
so just printf errors and use ErrorAbort at the very toplevel.

() read_write enforces that all read data is written; short writes
result in an error

() update_capsules() interface changed; now accepts a base directory
to look for capsules. If DMI machine props are in use, load all *.cap
capsules found in a subdirectory named by the machine-props file.
If not set or the base directory doesn't exist, don't do anything.

All the capsules are sent to the kernel unconditionally. There's no
specification for versioning, so it's up to the BIOS to make sure
updates aren't redundant or are downgraded. In addition no individual
verification of the capsules is done, although all update data comes
from or is patched by data within a signed container.

Change-Id: Iba2b3183bd38758a0198418d2d4d7b215fefe928
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://otc-android.intel.com/gerrit/19537
Tested-by: jenkins autobuilder
Works-for-me: jenkins autobuilder
Reviewed-by: Jianxun Zhang <jianxun.zhang@intel.com>
Reviewed-by: Michael Bergeron <michael.bergeron@intel.com>
Reviewed-by: Daniel Leung <daniel.leung@intel.com>