recovery: Implement new capsule update procedure
Vendors are treating the EFI spec as suggestions and it's not safe to
update capsules from the OS. Instead we put the capsules we want updated
on the EFI system partition and set a flag so that our bootloader does
it instead.

All the capsule files we want to install are copied from /system onto
the ESP since ext filesystems aren't supported. We recycle the logic which
selects device-appropriate capsules to install, if any. As the bootloader
pushes the capsules to be updated, the files are deleted.

We implement a 2-phase update; the second time we come into the Recovery
Console, we take action based on the BCB status field.

Change-Id: Iabba51f7d57cd9f9eebebbf7b12b0322b837eeda
Depends-Change-Id: I37f48407c37da7dd2aac1813d6839e5988bbcecf
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://otc-android.intel.com/gerrit/19607
Tested-by: jenkins autobuilder
Works-for-me: jenkins autobuilder
Works-for-me: Daniel Leung <daniel.leung@intel.com>
Reviewed-by: Michael Bergeron <michael.bergeron@intel.com>
Reviewed-by: Daniel Leung <daniel.leung@intel.com>